safari-600.7-branch/LayoutTests/fast/frames/sandboxed-iframe-navigation-top-denied.html

<html>
<head>
<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
</head>
<body>
<p>This test verifies that a sandboxed IFrame cannot navigate the top-level
frame without allow-top-navigation.  This test passes if the navigation does
not occur.</p>
<iframe sandbox="allow-scripts"
        src="resources/navigate-top-to-fail.html">
</body>
</html>
